requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.4
djangorestframework==3.14.0
Pillow==9.3.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
